+++ src/PatchCoverage.php
@@ -10,11 +10,10 @@
 namespace SebastianBergmann\PHPCOV;
 
 use const DIRECTORY_SEPARATOR;
-use function assert;
 use function file_get_contents;
 use function is_array;
 use function substr;
-use SebastianBergmann\CodeCoverage\CodeCoverage;
+use SebastianBergmann\CodeCoverage\Serialization\Unserializer;
 use SebastianBergmann\Diff\Line;
 use SebastianBergmann\Diff\Parser as DiffParser;
 
@@ -35,11 +34,8 @@ final class PatchCoverage
             $pathPrefix .= DIRECTORY_SEPARATOR;
         }
 
-        $coverage = include $coverageFile;
-
-        assert($coverage instanceof CodeCoverage);
-
-        $coverage = $coverage->getData()->lineCoverage();
+        $data     = (new Unserializer)->unserialize($coverageFile);
+        $coverage = $data['codeCoverage']->lineCoverage();
         $patch    = (new DiffParser)->parse(file_get_contents($patchFile));
         $changes  = [];