+++ src/cli/MergeCommand.php
@@ -13,7 +13,18 @@ use const PHP_EOL;
 use function is_dir;
 use function printf;
 use function realpath;
-use SebastianBergmann\CodeCoverage\CodeCoverage;
+use function serialize;
+use SebastianBergmann\CodeCoverage\Serialization\DriverMismatchException;
+use SebastianBergmann\CodeCoverage\Serialization\EmptyPathListException;
+use SebastianBergmann\CodeCoverage\Serialization\FileCouldNotBeReadException;
+use SebastianBergmann\CodeCoverage\Serialization\GitInformationMismatchException;
+use SebastianBergmann\CodeCoverage\Serialization\InvalidCoverageDataException;
+use SebastianBergmann\CodeCoverage\Serialization\Merger as CoverageMerger;
+use SebastianBergmann\CodeCoverage\Serialization\MixedGitInformationException;
+use SebastianBergmann\CodeCoverage\Serialization\RuntimeMismatchException;
+use SebastianBergmann\CodeCoverage\Serialization\VersionMismatchException;
+use SebastianBergmann\CodeCoverage\Util\Filesystem;
+use SebastianBergmann\CodeCoverage\Version;
 use SebastianBergmann\FileIterator\Facade;
 
 final class MergeCommand extends Command
@@ -49,44 +60,37 @@ final class MergeCommand extends Command
             return 1;
         }
 
-        $errors = [];
+        try {
+            $merged = (new CoverageMerger)->merge($files);
+        } catch (DriverMismatchException|RuntimeMismatchException|GitInformationMismatchException|MixedGitInformationException $e) {
+            print $e->getMessage() . PHP_EOL;
 
-        foreach ($files as $file) {
-            $_coverage = include $file;
+            return 1;
+        } catch (FileCouldNotBeReadException|VersionMismatchException|InvalidCoverageDataException $e) {
+            print 'Failed to merge: ' . $e->getMessage() . PHP_EOL;
 
-            if (!$_coverage instanceof CodeCoverage) {
-                $errors[] = $file;
-
-                unset($_coverage);
-
-                continue;
-            }
-
-            if (!isset($mergedCoverage)) {
-                $mergedCoverage = $_coverage;
-
-                continue;
-            }
-
-            $mergedCoverage->merge($_coverage);
-
-            unset($_coverage);
-        }
-
-        if (!isset($mergedCoverage)) {
-            foreach ($errors as $error) {
-                print 'Failed to merge: ' . $error . PHP_EOL;
-            }
-
+            return 1;
+        } catch (EmptyPathListException $e) {
             return 1;
         }
 
-        $this->handleReports($mergedCoverage, $arguments);
+        if ($arguments->php() !== null) {
+            print 'Generating code coverage report in PHP format ... ';
 
-        foreach ($errors as $error) {
-            print 'Failed to merge: ' . $error . PHP_EOL;
+            Filesystem::write(
+                $arguments->php(),
+                '<?php // phpunit/php-code-coverage version ' . Version::id() . PHP_EOL .
+                "return \unserialize(<<<'END_OF_COVERAGE_SERIALIZATION'" . PHP_EOL .
+                serialize($merged) . PHP_EOL .
+                'END_OF_COVERAGE_SERIALIZATION' . PHP_EOL .
+                ');',
+            );
+
+            print 'done' . PHP_EOL;
         }
 
-        return $errors === [] ? 0 : 1;
+        $this->handleReports($merged['codeCoverage'], $merged['testResults'], $arguments);
+
+        return 0;
     }
 }